% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervals.R
\name{add}
\alias{add}
\title{Add an interval to the map}
\usage{
add(x, start, end, value = NULL)
}
\arguments{
\item{x}{IntervalMap object}

\item{start}{Start position (inclusive)}

\item{end}{End position (inclusive)}

\item{value}{Associated value (optional)}
}
\value{
The IntervalMap object (invisibly)
}
\description{
Add an interval to the map
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
add(im, 15, 25, "gene2")
}
