% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervalsr.R
\name{clear}
\alias{clear}
\title{Clear all intervals}
\usage{
clear(x)
}
\arguments{
\item{x}{IntervalMap object}
}
\value{
The IntervalMap object (invisibly)

The IntervalMap object (invisibly)
}
\description{
Clear all intervals
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
clear(im)
}
