% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervals.R
\name{count}
\alias{count}
\title{Count overlapping intervals}
\usage{
count(x, start, end)
}
\arguments{
\item{x}{IntervalMap object}

\item{start}{Start of query range (inclusive)}

\item{end}{End of query range (inclusive)}
}
\value{
Number of overlapping intervals
}
\description{
Count overlapping intervals
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
add(im, 5, 15, "gene2")
build(im)
count(im, 7, 12)  # 2
}
