% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervalsr.R
\name{coverage}
\alias{coverage}
\title{Get coverage statistics for a range}
\usage{
coverage(x, start, end)
}
\arguments{
\item{x}{IntervalMap object}

\item{start}{Start of query range (inclusive)}

\item{end}{End of query range (inclusive)}
}
\value{
List with count and total_coverage
}
\description{
Get coverage statistics for a range
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
add(im, 5, 15, "gene2")
build(im)
coverage(im, 1, 20)
}
