% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervalsr.R
\name{data_at}
\alias{data_at}
\title{Get data at specific index}
\usage{
data_at(x, idx)
}
\arguments{
\item{x}{IntervalMap object}

\item{idx}{Index position (1-based)}
}
\value{
Associated data value

Associated data value at the specified index
}
\description{
Get data at specific index
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
build(im)
data_at(im, 1)  # Returns "gene1"
}
