% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervals.R
\name{length.IntervalMap}
\alias{length.IntervalMap}
\title{Get the length of an IntervalMap}
\usage{
\method{length}{IntervalMap}(x)
}
\arguments{
\item{x}{IntervalMap object}
}
\value{
Number of intervals in the map
}
\description{
Get the length of an IntervalMap
}
