% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervalsr.R
\name{reserve}
\alias{reserve}
\title{Reserve space for intervals}
\usage{
reserve(x, n)
}
\arguments{
\item{x}{IntervalMap object}

\item{n}{Number of intervals to reserve space for}
}
\value{
The IntervalMap object (invisibly)
}
\description{
Reserve space for intervals
}
