% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervalsr.R
\name{search_idxs}
\alias{search_idxs}
\title{Search for indices of overlapping intervals}
\usage{
search_idxs(x, start, end)
}
\arguments{
\item{x}{IntervalMap object}

\item{start}{Start of query range (inclusive)}

\item{end}{End of query range (inclusive)}
}
\value{
Integer vector of 1-based indices
}
\description{
Search for indices of overlapping intervals
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
add(im, 5, 15, "gene2")
build(im)
search_idxs(im, 7, 12)  # c(1, 2)
}
