% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervals.R
\name{search_items}
\alias{search_items}
\title{Search for complete items (start, end, value) of overlapping intervals}
\usage{
search_items(x, start, end)
}
\arguments{
\item{x}{IntervalMap object}

\item{start}{Start of query range (inclusive)}

\item{end}{End of query range (inclusive)}
}
\value{
List with start, end, and value vectors
}
\description{
Search for complete items (start, end, value) of overlapping intervals
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
add(im, 5, 15, "gene2")
build(im)
search_items(im, 7, 12)
}
