% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervals.R
\name{search_keys}
\alias{search_keys}
\title{Search for keys (start, end pairs) of overlapping intervals}
\usage{
search_keys(x, start, end)
}
\arguments{
\item{x}{IntervalMap object}

\item{start}{Start of query range (inclusive)}

\item{end}{End of query range (inclusive)}
}
\value{
List with start and end vectors
}
\description{
Search for keys (start, end pairs) of overlapping intervals
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
add(im, 5, 15, "gene2")
build(im)
search_keys(im, 7, 12)  # list(start=c(1,5), end=c(10,15))
}
