% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervalsr.R
\name{size}
\alias{size}
\title{Get number of intervals}
\usage{
size(x)
}
\arguments{
\item{x}{IntervalMap object}
}
\value{
Number of intervals

Integer number of intervals in the map
}
\description{
Get number of intervals
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
add(im, 20, 30, "gene2")
size(im)  # Returns 2
}
