% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervalsr.R
\name{starts_at}
\alias{starts_at}
\title{Get start position at specific index}
\usage{
starts_at(x, idx)
}
\arguments{
\item{x}{IntervalMap object}

\item{idx}{Index position (1-based)}
}
\value{
Start position

Integer start position of the interval
}
\description{
Get start position at specific index
}
\examples{
im <- IntervalMap()
add(im, 1, 10, "gene1")
build(im)
starts_at(im, 1)  # Returns 1
}
