% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervals.R
\name{print.IntervalMap}
\alias{print.IntervalMap}
\title{Print method for IntervalMap}
\usage{
\method{print}{IntervalMap}(x, ...)
}
\arguments{
\item{x}{IntervalMap object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for IntervalMap
}
