% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintervals.R
\name{starts_at}
\alias{starts_at}
\title{Get start position at specific index}
\usage{
starts_at(x, idx)
}
\arguments{
\item{x}{IntervalMap object}

\item{idx}{Index position (1-based)}
}
\value{
Start position
}
\description{
Get start position at specific index
}
